LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/sandboxing-back-parent-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to initiate navigation for frame with URL 'http://localhost:8800/navigation-api/navigation-methods/sandboxing-back-parent.html#end' from frame with URL 'http://localhost:8800/navigation-api/navigation-methods/resources/navigation-back.html'. The frame attempting navigation of the top-level window is sandboxed, but the 'allow-top-navigation' flag is not set.



PASS A sandboxed iframe cannot navigate its parent via its own navigation object by using back()
